% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comp_expmat}
\alias{comp_expmat}
\title{Compute the matrix exponential.}
\usage{
comp_expmat(Q)
}
\arguments{
\item{Q}{matrix}
}
\value{
Matrix exponential of Q
}
\description{
Compute the matrix exponential.
}
