% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_path_unif}
\alias{sample_path_unif}
\title{Simulate a sample path from an endpoint conditioned CTMC by uniformization.}
\usage{
sample_path_unif(a, b, t0, t1, Q)
}
\arguments{
\item{a, b}{States at the interval endpoints, provided as integers
corresponding to rows of the CTMC rate matrix.}

\item{t0, t1}{times of the interval endpoints}

\item{Q}{CTMC rate matrix}
}
\value{
matrix whose first column is the sequence of transition times
bookended by interval endpoints, and whose second column is the sequence of
states
}
\description{
Simulate a sample path from an endpoint conditioned CTMC by uniformization.
}
